pyproject.toml
[project]
name = "waldo"
version = "0.1.0"
description = "SAM 3 + YOLO auto-labeling pipeline"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0",
    "alembic",
    "psycopg2-binary",
    "redis",
    "minio",
    "celery[redis]",
    "pydantic>=2.0",
    "pydantic-settings",
]

[dependency-groups]
app = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
]
labeler = [
    "transformers>=4.47",
    "torch>=2.5",
    "Pillow",
    "opencv-python-headless",
    "imagehash",
    "numpy",
    "torchvision>=0.25.0",
    "kernels>=0.12.2",
]
trainer = [
    "ultralytics",
    "lapx",
    "requests",
]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lib", "app", "labeler", "trainer"]

[tool.uv]
default-groups = ["app", "labeler", "trainer", "dev"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
